# microservices/order-management/src/main/openapi/product-management-0.0.yml
openapi: 3.0.1
info:
  title: product-management
  version: "0.0"
paths:
  /ordersByDays:
    get:
      tags:
        - ordersByDays
      operationId: getOrdersByDays
      responses:
        "200":
          description: getOrdersByDays 200 response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrdersByDay"
  /ordersByDays/day/{day}:
    get:
      tags:
        - ordersByDays
      operationId: getByProduct
      parameters:
        - name: day
          in: path
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: getByProduct 200 response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrdersByDay"
  /ordersByDays/day/{day}/product/{productId}:
    get:
      tags:
        - ordersByDays
      operationId: getByProductAndDate
      parameters:
        - name: day
          in: path
          required: true
          schema:
            type: string
            format: date
        - name: productId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: getByProductAndDate 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrdersByDay"
  /ordersByDays/id/{id}:
    get:
      tags:
        - ordersByDays
      operationId: getOrdersByDay
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: getOrdersByDay 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrdersByDay"
    delete:
      tags:
        - ordersByDays
      operationId: deleteOrdersByDay
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: deleteOrdersByDay 200 response
  /ordersByDays/product/{productId}:
    get:
      tags:
        - ordersByDays
      operationId: getByProduct_1
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: getByProduct_1 200 response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrdersByDay"
  /products:
    get:
      tags:
        - products
      operationId: getProducts
      responses:
        "200":
          description: getProducts 200 response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      tags:
        - products
      operationId: createProduct
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductCreateDTO"
        required: true
      responses:
        "200":
          description: createProduct 200 response
  /products/{id}:
    get:
      tags:
        - products
      operationId: getProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: getProduct 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    put:
      tags:
        - products
      operationId: updateProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductCreateDTO"
        required: true
      responses:
        "200":
          description: updateProduct 200 response
    delete:
      tags:
        - products
      operationId: deleteProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: deleteProduct 200 response
  /products/{id}/price:
    get:
      tags:
        - products
      operationId: getProductPrice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: getProductPrice 200 response
          content:
            application/json:
              schema:
                type: number
  /products/{id}/tags:
    get:
      tags:
        - products
      operationId: getProductTags
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: getProductTags 200 response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tag"
  /products/{id}/tags/{tagId}:
    put:
      tags:
        - products
      operationId: addProductTag
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: tagId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: addProductTag 200 response
    delete:
      tags:
        - products
      operationId: removeProductTag
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: tagId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: removeProductTag 200 response
  /tags:
    get:
      tags:
        - tags
      operationId: getTags
      responses:
        "200":
          description: getTags 200 response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tag"
    post:
      tags:
        - tags
      operationId: createTag
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagCreateDTO"
        required: true
      responses:
        "200":
          description: createTag 200 response
  /tags/{id}:
    get:
      tags:
        - tags
      operationId: getTag
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: getTag 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
    put:
      tags:
        - tags
      operationId: updateTag
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagCreateDTO"
        required: true
      responses:
        "200":
          description: updateTag 200 response
    delete:
      tags:
        - tags
      operationId: deleteTag
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: deleteTag 200 response
components:
  schemas:
    OrdersByDay:
      type: object
      properties:
        id:
          type: integer
          format: int64
          nullable: true
        day:
          type: string
          format: date
        count:
          type: integer
          format: int32
    Product:
      type: object
      properties:
        id:
          type: integer
          format: int64
          nullable: true
        name:
          type: string
        unitPrice:
          type: number
    ProductCreateDTO:
      type: object
      properties:
        name:
          type: string
        unitPrice:
          type: number
    Tag:
      type: object
      properties:
        id:
          type: integer
          format: int64
          nullable: true
        name:
          type: string
    TagCreateDTO:
      type: object
      properties:
        name:
          type: string